Feat: 통합광고현황 컴포넌트 분리, 계산식 수정 및 정리
