+++ src/routes/DashBoard/IntegratedAdManagement/IntergratedAdStatus/intergratedAdStatus.module.scss
@@ -1,0 +1,6 @@
+.group {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  padding: 40px;
+}
