+++ src/routes/DashBoard/IntegratedAdManagement/IntergratedAdStatus/status.ts
@@ -1,0 +1,32 @@
+export const chartOptions = [
+  {
+    value: 'roas',
+    content: 'ROAS',
+    unit: '%',
+  },
+  {
+    value: 'cost',
+    content: '광고비',
+    unit: '원',
+  },
+  {
+    value: 'cpc',
+    content: '노출 수',
+    unit: '회',
+  },
+  {
+    value: 'click',
+    content: '클릭 수',
+    unit: '회',
+  },
+  {
+    value: 'conv',
+    content: '전환 수',
+    unit: '회',
+  },
+  {
+    value: 'convValue',
+    content: '매출',
+    unit: '원',
+  },
+]
